backend/requirements.txt
paho-mqtt
mysql-connector-python
twilio
Flask
plotly